# lib/api-zod/src/generated/types/mobileMemoryCreateRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchSourceMetadata } from './researchSourceMetadata';

export interface MobileMemoryCreateRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  clientId: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 8000
     */
  finding: string;
  entities?: string[];
  sources?: ResearchSourceMetadata[];
}