# lib/api-zod/src/generated/types/mobileMemoryResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MobileMemoryResult {
  id: string;
  title: string;
  finding: string;
  savedAt: Date;
  sourceCount: number;
}