# lib/api-zod/src/generated/types/mobileProjectCreateRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MobileProjectCreateRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  clientId: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /** @maxLength 1000 */
  description?: string;
}