# lib/api-zod/src/generated/types/mobileProjectResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchHistorySummary } from './researchHistorySummary';

export interface MobileProjectResult {
  id: string;
  name: string;
  description: string;
  analysisCount: number;
  sourceCount: number;
  lastActivityAt: Date;
  createdAt: Date;
  updatedAt: Date;
  recentResearch: ResearchHistorySummary[];
}